+++ src/yabf/__init__.py
@@ -7,7 +7,9 @@ try:
 except PackageNotFoundError:
     __version__ = "unknown"
 
-from .core import mpi, samplers
+from . import samplers
+from .core import mpi
+from .core._samplers import Sampler, curve_fit, run_map
 from .core.configio import (
     load_from_yaml,
     load_likelihood_from_yaml,
@@ -15,4 +17,3 @@ from .core.configio import (
 )
 from .core.likelihood import Component, Likelihood, LikelihoodContainer
 from .core.parameters import Param, Parameter, ParameterVector, ParamVec
-from .core.samplers import run_map
