+++ src/yabf/core/__init__.py
@@ -1,3 +1,3 @@
 """Core functionality for YABF."""
 
-from . import component, configio, io, likelihood, mpi, parameters, samplers, utils
+from . import _samplers, component, configio, io, likelihood, mpi, parameters, utils
